requirements.txt
Flask==3.0.3
pyttsx3==2.90
wikipedia-api==0.5.4
